Moved `LoadedLibraryData` inside of `NativeLibrary` because it fits better inside it
